Sort method parameters
Sort method parameters so that the order of the parameters is always the
same (so long as the method interface does not change). For APIs the
parameters are sorted by name. For model constructors the "position"
attribute of a property's definition is used to determine its position
in the parameter list. Properties with the same position value are
sorted by name. Of course, mandatory parameters always appear before
optional parameters.

Imports are also sorted by module name.