Typecast response codes to string before comparing
The PingFederate swagger specification used to define response codes as
integers when using Swagger 1.2 but defines them as strings in Swagger
2.0. Therefore, to preserve compatibility with older PingFederate
versions, the code templates should typecast response codes to a
common (known) type before making comparisons (to decide on what code
needs to be emitted for what response types).

Note the actual response codes during execution of the generated code
will always be integers - the response codes being processed by the
generator are the response codes specified in the API Swagger
definitions.