Schema patches for some models
The Swagger schema for the following models need patching:
    - ArtifactSettings
    - AttributeFulfillmentValue
    - SpBrowserSso
    - SpBrowserSsoAttribute
    - SpSsoServiceEndpoint
    - SpWsTrustAttribute

Details about the reasons are in the docstring in the respective patch
files.

With these changes the manually applied (and incorrect) changes to the
generated SDK in the following commits should be removed:
    - a962972f6b54626f7b345513905dd44ef301dc4b
    - 56655675b042f5b241fe9bb370a043e1d91b78dc

The fetcher has also been changed to save patched Swagger schemas so
that it is easy to compare the state of the Swagger schemas before and
after patching.

Also fix failing unit test.